frontend-otel/requirements.txt
Flask==2.3.3
requests==2.31.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-semantic-conventions==0.41b0
protobuf==4.24.3
grpcio==1.58.0
opentelemetry-exporter-otlp==1.20.0
